requirements.txt
# app
fastapi==0.136.0
pydantic==2.13.3
pydantic-settings==2.14.0
uvicorn[standard]==0.45.0
apscheduler[mongodb]==4.0.0a6

# http client
httpx==0.28.1

# mongodb
pymongo==4.17.0

# conversion
markitdown[all]==0.1.5
docling==2.90.0
torch==2.11.0
torchvision==0.26.0
pymupdf==1.27.2.2

# LLM providers (for ACCURATE strategy)
openai==2.32.0
anthropic==0.96.0
google-genai==1.73.1

# google drive auth (for service account)
PyJWT==2.12.1
cryptography==46.0.7

# sftp
paramiko==4.0.0

# azure blob storage
azure-storage-blob==12.28.0
aiohttp==3.13.5

# aws s3
aiobotocore==3.5.0

# google cloud storage
google-cloud-storage==3.10.1

# dev
python-dotenv==1.2.2
